requirements.txt
pywebview~=4.0.2
nicegui~=1.1.11
uvicorn~=0.20.0